fix(fs): fix NotFound error when moving src to itself with overwrite: true (#3289)

+++ fs/move.ts
@@ -1,7 +1,15 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-import { isSubdir } from "./_util.ts";
+import { isSamePath, isSubdir } from "./_util.ts";
 
 const EXISTS_ERROR = new Deno.errors.AlreadyExists("dest already exists.");
+
+export class SubdirectoryMoveError extends Error {
+  constructor(src: string | URL, dest: string | URL) {
+    super(
+      `Cannot move '${src}' to a subdirectory of itself, '${dest}'.`,
+    );
+  }
+}
 
 interface MoveOptions {
   overwrite?: boolean;
@@ -24,13 +32,15 @@ export async function move(
 ) {
   const srcStat = await Deno.stat(src);
 
-  if (srcStat.isDirectory && isSubdir(src, dest)) {
-    throw new Error(
-      `Cannot move '${src}' to a subdirectory of itself, '${dest}'.`,
-    );
+  if (
+    srcStat.isDirectory &&
+    (isSubdir(src, dest) || isSamePath(src, dest))
+  ) {
+    throw new SubdirectoryMoveError(src, dest);
   }
 
   if (overwrite) {
+    if (isSamePath(src, dest)) return;
     try {
       await Deno.remove(dest, { recursive: true });
     } catch (error) {
@@ -66,13 +76,15 @@ export function moveSync(
 ) {
   const srcStat = Deno.statSync(src);
 
-  if (srcStat.isDirectory && isSubdir(src, dest)) {
-    throw new Error(
-      `Cannot move '${src}' to a subdirectory of itself, '${dest}'.`,
-    );
+  if (
+    srcStat.isDirectory &&
+    (isSubdir(src, dest) || isSamePath(src, dest))
+  ) {
+    throw new SubdirectoryMoveError(src, dest);
   }
 
   if (overwrite) {
+    if (isSamePath(src, dest)) return;
     try {
       Deno.removeSync(dest, { recursive: true });
     } catch (error) {
